+++ packages/terra-table/src/subcomponents/ColumnHeader.jsx
@@ -97,6 +97,7 @@ const ColumnHeader = (props) => {
             rowIndex={0}
             columnIndex={columnIndex}
             displayName={column.displayName}
+            isDisplayVisible={column.isDisplayVisible}
             width={column.width}
             minimumWidth={column.minimumWidth}
             maximumWidth={column.maximumWidth}