+++ packages/terra-data-grid/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed
   * Fixed issue where focus was given to the column header instead of its button element.
+  * Fixed issue where row selection was being announced twice in Worklist Data Grid.
   * Fixed an issue when data accesses invalid indexes when deleting rows and columns.
 
 * Added